+++ Changelog.md
@@ -6,6 +6,9 @@
 
 ### Ignition Rendering 4.0.0
 
+1. Skeleton animation updates
+    * [Pull request 235](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/235)
+
 1. Add support for transparency based on textures alpha channel for ogre1 and ogre2
     * [Pull request 229](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/229)